+++ src/Service/StylesheetProvider.php
@@ -1,0 +1,22 @@
+<?php
+/*
+ * (c) 2026: 975L <contact@975l.com>
+ * (c) 2026: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+namespace c975L\SiteBundle\Service;
+
+use c975L\UiBundle\Contract\BundleStylesheetProviderInterface;
+
+class StylesheetProvider implements BundleStylesheetProviderInterface
+{
+    public function getStylesheets(): array
+    {
+        return [
+            'bundles/c975lsite/css/styles.min.css',
+            'https://cdnjs.cloudflare.com/ajax/libs/cookieconsent2/3.1.0/cookieconsent.min.css',
+        ];
+    }
+}